Move event share into dedicated file
